internal/cli/internal/version: split to file per sub-command
